Improved encapsulation of the Memento pattern by moving the StarMementoInternal class to inner class of Star.
