+++ memento/src/main/java/com/iluwatar/Star.java
@@ -66,4 +66,40 @@ public class Star {
 		return String.format("%s age: %d years mass: %d tons", type.toString(),
 				ageYears, massTons);
 	}
+	
+	/**
+	 * 
+	 * StarMemento implementation
+	 * 
+	 */
+	private static class StarMementoInternal implements StarMemento {
+
+		private StarType type;
+		private int ageYears;
+		private int massTons;
+
+		public StarType getType() {
+			return type;
+		}
+
+		public void setType(StarType type) {
+			this.type = type;
+		}
+
+		public int getAgeYears() {
+			return ageYears;
+		}
+
+		public void setAgeYears(int ageYears) {
+			this.ageYears = ageYears;
+		}
+
+		public int getMassTons() {
+			return massTons;
+		}
+
+		public void setMassTons(int massTons) {
+			this.massTons = massTons;
+		}
+	}
 }
